+++ example/src/home/pages/detail/effects/index.js
@@ -1,20 +1,17 @@
-import BaseEffects from '../../../../public/effects';
 import services from '../services';
 
-export default class Effects extends BaseEffects {
+export default {
   count() {
     const { count } = this.getState();
     this.updateState({
       count: count + 1,
     });
-  }
-
+  },
   async $getDetail() {
     const data = await services.getDetail();
     this.updateState(data);
-  }
-
+  },
   async initData() {
     await this.$getDetail();
-  }
-}
+  },
+};