+++ example/src/home/pages/detail/index.js
@@ -1,34 +1,13 @@
 import React from 'react';
-import Effects from './effects';
+import effects from './effects';
 import Layout from './components/Layout';
-import services from './services';
 
 export default {
   state: {
     detail: '',
     count: 0,
   },
-  effects: {
-    updateState(payload) {
-      this.dispatch({
-        type: '_updateState',
-        payload,
-      });
-    },
-    count() {
-      const { count } = this.getState();
-      this.updateState({
-        count: count + 1,
-      });
-    },
-    async $getDetail() {
-      const data = await services.getDetail();
-      this.updateState(data);
-    },
-    async initData() {
-      await this.$getDetail();
-    },
-  },
+  effects,
   render() {
     return <Layout />;
   },