+++ example/src/public/config.js
@@ -8,4 +8,12 @@ if (process.env.NODE_ENV !== 'production') {
 
 nuomi.config({
   wrapper: true,
+  effects: {
+    updateState(payload) {
+      this.dispatch({
+        type: '_updateState',
+        payload,
+      });
+    },
+  },
 });