+++ src/components/Route/index.jsx
@@ -4,8 +4,17 @@ import invariant from 'invariant';
 import RouterContext from '../RouterContext';
 import RouteCore from '../RouteCore';
 import { removeReducer } from '../../core/redux/reducer';
-import { matchPath, savePath, removePath, getParams } from '../../core/router';
+import {
+  matchPath,
+  savePath,
+  removePath,
+  getParamsLocation,
+  normalLocation,
+  location as routerLocation,
+} from '../../core/router';
 import { getDefaultProps } from '../../core/nuomi';
+
+let activeRouteComponent = null;
 
 class Route extends React.PureComponent {
   static propTypes = {
@@ -43,6 +52,7 @@ class Route extends React.PureComponent {
     this.store = {};
     this.routeTempData = {};
     this.ref = React.createRef();
+    this.routeComponent = null;
     const { path } = this.props;
     savePath(path);
   }
@@ -64,6 +74,9 @@ class Route extends React.PureComponent {
         current.ref.current.removeListener();
       }
     }
+    if (activeRouteComponent === this.routeComponent) {
+      activeRouteComponent = null;
+    }
     removePath(path);
     removeReducer(this.store.id);
   }
@@ -82,25 +95,65 @@ class Route extends React.PureComponent {
             match = false;
           }
           // 设置了wrapper没有匹配路由，不销毁，只隐藏
-          if (wrapper === true && this.routeCore && !match) {
-            return this.routeCore;
+          if (wrapper === true && this.routeComponent !== null && !match) {
+            return this.routeComponent;
           }
+          // 还原路由时，不重新渲染组件
+          if (context.restore) {
+            return this.routeComponent;
+          }
+          // 检测之前的路由onLeave
+          if (
+            !!activeRouteComponent &&
+            !!activeRouteComponent.ref.current &&
+            !context.callOnLeave
+          ) {
+            const baseRouteComponent = activeRouteComponent.ref.current.ref.current;
+            if (baseRouteComponent) {
+              const { props } = baseRouteComponent;
+              if (props.onLeave) {
+                const leave = () => {
+                  // 1.防止跳转后再次执行onLeave导致死循环，2.用作调用leave后的标记
+                  activeRouteComponent = null;
+                  routerLocation(location, location.data, location.reload);
+                };
+                const leaveResult = props.onLeave(() => leave());
+                if (activeRouteComponent === null) {
+                  invariant(
+                    false,
+                    'onLeave中不能直接进行路由跳转，可以通过返回布尔值控制，' +
+                      '如果想确认框确认或者异步操作之后跳转，请将逻辑代码置于return false之前，在回调中调用leave方法。',
+                  );
+                }
+                // 防止onLeave重复执行
+                context.callOnLeave = true;
+                if (leaveResult === false) {
+                  // 还原路由标记
+                  context.restore = true;
+                  // 还原为之前的路由，还原时所有的监听不会执行
+                  normalLocation(props.location);
+                  return this.routeComponent;
+                }
+              }
+            }
+          }
+          // 初始化返回值
+          this.routeComponent = null;
           if (match) {
-            // eslint-disable-next-line no-param-reassign
-            context.matched = this; // 解决Route在更新时不匹配问题，值不能设置为true
-            location.params = getParams(location, path); // 解析动态参数
-            this.routeCore = (
+            context.matched = this; // 解决Route在更新时不匹配问题
+            // 记录当前路由
+            activeRouteComponent = (
               <RouteCore
                 {...this.props}
                 wrapper={wrapper}
-                location={location}
+                location={getParamsLocation(location, path)}
                 store={this.store}
                 ref={this.ref}
               />
             );
-            return this.routeCore;
+            this.routeComponent = activeRouteComponent;
           }
-          return null;
+          return this.routeComponent;
         }}
       </RouterContext.Consumer>
     );